requirements.txt
asgiref==3.6.0
boto3==1.26.145
botocore==1.29.145
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.1.7
django-extensions==3.2.1
django-heroku==0.3.1
django-storages==1.13.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
jmespath==1.0.1
Pillow==9.4.0
psycopg2==2.9.6
python-dateutil==2.1
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.16
whitenoise==6.4.0
